Behat test suite (#70)
* prepare behat for test suite

* add FacultyFixtures

* add behat.php to gitignore

* remove unused object-manager

* Add AuthContext Behat test

* mago: use value-of<T> for enum values validation

* fmt

* perf: use cow in results

* activity is not nullable

* fix: we don't use faculty.parent_id anymore

* mago: indent-binary-expression-continuation

* remove unused namespace variable

* add user fixtures, init database for in memory db, perf

* fix mago issues, use in-memory-db only for now

* fix: use `array_key_first` instead of `key`

`array_key_first` actually gives the first key, `key` could return
gibberish as the internal pointer could point to some other element,
although it is impossible in this context, because we're checking for
the `count($arr)`...

* mago: add more checks

* fix: SubmissionStateDto will now correctly accept booleans

* just: add justfile

* just: fix command naming

* format mago workflow file

* add tests workflow file

* ci: add just to tests.yml

* ci: add APP_ENV=test to run in test environment

* tests: add login test

* fix: submission accept/reject actions

* profile cache fixer now fixes all users

+++ src/Action/SubmissionActions.php
@@ -115,11 +115,20 @@ final readonly class SubmissionActions
             return ['mismatch_updated_at'];
         }
 
+        // Submission is reviewed and DTO has the same state as the submission => noop
+        // TODO(@bleksak): The message might have changed, but we will handle that another day
+        if (
+            $submission->reviewed === true
+            && $submission->accepted === $dto->state
+        ) {
+            return [];
+        }
+
         $submission->message = $dto->message;
 
         $this->handleCacheUpdate($submission, $dto);
 
-        // TODO(@bleksak): Ted mi doslo, ze tady se da frajerovi zaspamovat email kdyby kutak furt schvaloval a zamital aktivitu :D
+        // TODO(@bleksak): tady se da frajerovi zaspamovat email kdyby admin furt schvaloval a zamital aktivitu :D
         if (!$dto->state) {
             // if ($submission->getUser()->getToken() !== null) {
             //     $this->firebase->send(new VytrvalecNotification($submission->getUser(), $dto->message));
@@ -226,12 +235,12 @@ final readonly class SubmissionActions
         Submission $submission,
         SubmissionStateDto $dto,
     ): void {
-        if ($dto->state && (!$submission->reviewed || !$submission->accepted)) {
+        if ($dto->state && !$submission->reviewed && !$submission->accepted) {
             // noop when already accepted, otherwise profile cache would stack
             $this->profileCacheRepository->addCache($submission);
         }
 
-        if (!$dto->state) {
+        if (!$dto->state && $submission->accepted) {
             $this->profileCacheRepository->removeCache($submission);
         }
     }
